Create README.md
LoveRunningCI Introduction This is a website designed to create interest in running based on a run club in Dublin Docklands and Pheonix Park. The design is focused on including basic information of how to sign up, Where you will find us, and a gallery showing people enjoying running.

A fully responsive website using CSS and HTML5 Used to apply and demonstrate the HTML and CSS elements and concepts Used to apply and demonstrate the HTML/CSS3 elements and concepts covered in Code Institute's Fundamentals Module Series

Acknowledgements (Code Institute) This project is part of our digital learning for code institute- all images and designs were provided by Code Institute.